// components/model/products/ProductsData.tsx
export interface InsuranceProduct {
  id: number;
  period_name: string;
  insurance_type_id: number;
  insurance_category: string;
  premium_percentage: number;
  insurance_period_id: number;
  insurance_type_name: string;
  insurance_company_id: number;
  insurance_category_id: number;
  insurance_company_name: string;
  insurance_company_user_id: number;
}